Refactor: RandomBallsGenerator를 유틸이 아닌 일반 클래스로 변경
