+++ src/main/java/baseball/domain/balls/RandomBallsGenerator.java
@@ -1,17 +1,18 @@
-package baseball.utility;
+package baseball.domain.balls;
 
 import baseball.domain.balls.Balls;
+import baseball.utility.ConvertorUtils;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import java.util.stream.Stream;
 
-public final class RandomBallsGenerator {
+public class RandomBallsGenerator {
 
     private static final int MIN_VALUE = 1;
     private static final int MAX_VALUE = 9;
     private static final int BALL_COUNT = 3;
 
-    public static Balls generate() {
+    public Balls generate() {
         List<Integer> randomIntegers = Stream.generate(() -> generateRandomNumber())
                 .distinct()
                 .limit(BALL_COUNT)
@@ -19,7 +20,7 @@ public final class RandomBallsGenerator {
         return ConvertorUtils.convertIntegersToBalls(randomIntegers);
     }
 
-    private static int generateRandomNumber() {
+    private int generateRandomNumber() {
         return Randoms.pickNumberInRange(MIN_VALUE, MAX_VALUE);
     }
